tests/wpt/web-platform-tests/portals/portals-nested.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <script>
    promise_test(() => {
      assert_implements("HTMLPortalElement" in self);
      var portal = document.createElement("portal");
      portal.src = "resources/portals-nested-portal.html";
      document.body.appendChild(portal);
      var waitForMessage = new Promise((resolve, reject) => {
        portal.onmessage = resolve;
      });
      return waitForMessage;
    }, "nested portals shouldn't crash");
  </script>
</body>
